Move all functions from utils to helper
