+++ tests/unit/test_datatypes.py
@@ -1,36 +1,15 @@
 """Tests for the xmigrate.datatypes module."""
 
-import unittest.mock
-
 import pytest
 
 import xmigrate
-
-
-def _make_connection(datatypes: list[str]) -> unittest.mock.MagicMock:
-    """
-    Create a mock XNAT connection returning the given datatype element names.
-
-    Parameters
-    ----------
-    datatypes
-        The list of datatype element names to be returned by the mock connection.
-
-
-    Returns
-    -------
-        A mock XNAT connection.
-
-    """
-    conn = unittest.mock.MagicMock(spec_set=["get"])
-    conn.get.return_value.json.return_value = [{"elementName": dt} for dt in datatypes]
-    return conn
+from tests.fixtures.helpers import make_connection
 
 
 def test_check_datatypes_matching_identical() -> None:
     """No error raised when source and destination have identical datatypes."""
-    source = _make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
-    destination = _make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
+    source = make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
+    destination = make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
     xmigrate.check_datatypes_matching(source, destination)
     source.get.assert_called_once_with("/xapi/access/displays/createable")
     destination.get.assert_called_once_with("/xapi/access/displays/createable")
@@ -38,8 +17,8 @@ def test_check_datatypes_matching_identical() -> None:
 
 def test_check_datatypes_matching_destination_superset() -> None:
     """No error raised when destination has more datatypes than source."""
-    source = _make_connection(["xnat:mrSessionData"])
-    destination = _make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
+    source = make_connection(["xnat:mrSessionData"])
+    destination = make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
     xmigrate.check_datatypes_matching(source, destination)
     source.get.assert_called_once_with("/xapi/access/displays/createable")
     destination.get.assert_called_once_with("/xapi/access/displays/createable")
@@ -47,8 +26,8 @@ def test_check_datatypes_matching_destination_superset() -> None:
 
 def test_check_datatypes_matching_missing_on_destination() -> None:
     """ValueError raised when source has datatypes not enabled on destination."""
-    source = _make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
-    destination = _make_connection(["xnat:mrSessionData"])
+    source = make_connection(["xnat:mrSessionData", "xnat:ctSessionData"])
+    destination = make_connection(["xnat:mrSessionData"])
     with pytest.raises(
         ValueError,
         match=r"Source has datatypes not enabled on destination: {'xnat:ctSessionData'}",
@@ -60,8 +39,8 @@ def test_check_datatypes_matching_missing_on_destination() -> None:
 
 def test_check_datatypes_matching_xdat_filtered_out() -> None:
     """xdat: prefixed datatypes are excluded from the comparison."""
-    source = _make_connection(["xnat:mrSessionData", "xdat:something"])
-    destination = _make_connection(["xnat:mrSessionData"])
+    source = make_connection(["xnat:mrSessionData", "xdat:something"])
+    destination = make_connection(["xnat:mrSessionData"])
     xmigrate.check_datatypes_matching(source, destination)
     source.get.assert_called_once_with("/xapi/access/displays/createable")
     destination.get.assert_called_once_with("/xapi/access/displays/createable")
@@ -69,8 +48,8 @@ def test_check_datatypes_matching_xdat_filtered_out() -> None:
 
 def test_check_datatypes_matching_empty_source() -> None:
     """No error raised when source has no datatypes."""
-    source = _make_connection([])
-    destination = _make_connection(["xnat:mrSessionData"])
+    source = make_connection([])
+    destination = make_connection(["xnat:mrSessionData"])
     xmigrate.check_datatypes_matching(source, destination)
     source.get.assert_called_once_with("/xapi/access/displays/createable")
     destination.get.assert_called_once_with("/xapi/access/displays/createable")
